requirements.txt
absl-py==1.3.0
aes-cipher==2.0.0
amqp==5.0.6
antlr4-python2-runtime==4.10.1
antlr4-python3-runtime==4.10
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astgen==0.1.6.13
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==21.12b0
bleach==1.5.0
bs4==0.0.1
cachetools==5.2.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.4
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.8
django-cache-url==3.2.3
django-celery-beat==2.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
django-dbbackup==3.3.0
django-filter==21.1
django-heroku==0.3.1
django-jalali-date==1.0.0
django-jenkins==0.110.0
django-property-filter==1.1.1
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dl-coursera==0.1.2
drf-nested-routers==0.93.4
drf-yasg==1.17.1
entrypoints==0.3
environs==8.0.0
exceptiongroup==1.2.2
executing==0.8.2
fastjsonschema==2.16.2
filelock==3.4.0
flatbuffers==22.9.24
flower==1.0.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
gunicorn==20.1.0
h5py==3.7.0
html5lib==0.9999999
humanize==3.13.1
idna==3.3
imageio==2.22.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==8.0.2
itypes==1.2.0
jdatetime==3.6.4
jedi==0.18.1
Jinja2==3.0.3
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.4
jupyter-core==4.9.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-visualizer==2.4
kiwisolver==1.3.2
kombu==5.2.2
libclang==14.0.6
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.1.1
marshmallow==3.14.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mistune==2.0.4
mock==4.0.3
mpmath==1.2.1
mypy-extensions==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.4
networkx==2.8.7
notebook==6.4.12
numpy==1.22.1
oauthlib==3.1.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pddlpy==0.3.3
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2021.11.23
pkg-about==1.0.3
pkgutil_resolve_name==1.3.10
platformdirs==2.4.0
pluggy==1.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.6
psycopg2==2.9.2
psycopg2-binary==2.8.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.17
pygame==2.1.0
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.6
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
pytesseract==0.3.10
pytest==8.3.3
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyWavelets==1.4.1
pywin32==303
pywinpty==2.0.8
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.2
QtPy==2.2.0
redis==5.2.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.9
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-image==0.19.3
scikit-learn==1.1.3
scikit-surprise==1.1.4
scipy==1.9.3
Send2Trash==1.8.0
simpful==2.9.0
six==1.16.0
skyroom==1.0.3
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.2
stack-data==0.1.4
sympy==1.8
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.3
torch==1.13.0
tornado==6.1
tqdm==4.64.1
traitlets==5.4.0
typing-extensions==3.10.0.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.7
uuid==1.30
validate-email==1.3
vine==5.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wget==3.2
whitenoise==5.3.0
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.6.0
